broadCastMessage utils
